Move card to matching column
