+++ __mocks__/node-fetch.js
@@ -1,11 +1,12 @@
 /* eslint-env jest */
 // eslint-disable-next-line import/no-extraneous-dependencies
 const { Response, Headers } = jest.requireActual('node-fetch');
-const getProjectColumns = require('../test/fixtures/github/project.columns.json');
-const getProjectCard = require('../test/fixtures/github/project.card.json');
-const getProjectCardNoZube = require('../test/fixtures/github/project.card.no.zube.json');
+const getProjectColumns = require('../test/fixtures/github/project-columns.json');
+const getProjectCard = require('../test/fixtures/github/project-card.json');
+const getProjectCardNoZube = require('../test/fixtures/github/project-card-no-zube.json');
+const getProjectCardDiffZube = require('../test/fixtures/github/project-card-diff-zube.json');
 const getProject = require('../test/fixtures/github/project.json');
-const getProjectNoCards = require('../test/fixtures/github/project.no.cards.json');
+const getProjectNoCards = require('../test/fixtures/github/project-no-cards.json');
 const getLabel = require('../test/fixtures/github/label.json');
 
 const fixtures = {
@@ -13,6 +14,7 @@ const fixtures = {
   'query getProjectCard': {
     1: getProjectCardNoZube,
     2: getProjectCard,
+    3: getProjectCardDiffZube,
   },
   'query getProjectFromIssue': {
     1: getProject,
