+++ test/fixtures/github/project-card-created.json
@@ -1,0 +1,14 @@
+{
+    "action": "created",
+    "project_card": {
+        "node_id": "MDExOlByb2plY3RDYXJkMzkyMDc2MTE="
+    },
+    "repository": {
+        "id": 186853002,
+        "node_id": "MDEwOlJlcG9zaXRvcnkxODY4NTMwMDI=",
+        "name": "Hello-World",
+        "owner": {
+          "login": "Codertocat"
+        }
+    }
+}
