--- test/fixtures/github/project_card.created.json
@@ -1,6 +1,0 @@
-{
-    "action": "created",
-    "project_card": {
-        "node_id": "MDExOlByb2plY3RDYXJkMzkyMDc2MTE="
-    }
-}
